Cleanup
git-svn-id: http://equinox:10080/_svn_root/AObjectServer@271 1a28b8b1-d4e8-a449-a3f8-e3d44feecac6

+++ AObjectServer/AOS_Base/AOSContextQueueThreadPool_RoundRobinSwarm.cpp
@@ -65,7 +65,7 @@ AOSContextQueueThreadPool_RoundRobinSwarm::AOSContextQueueThreadPool_RoundRobinS
   m_queueCount(queueCount),
   AOSContextQueueThreadPool(services, threadCount),
   m_currentWriteQueue(0),
-  m_currentReadQueue(queueCount-1)
+  m_currentReadQueue(((long)queueCount)-1)
 {
   AASSERT(this, queueCount > 0);
   m_Queues.resize(m_queueCount);
